oxish: move Server::spawn() into unix
